+++ src/types.ts
@@ -9,7 +9,7 @@ export type JsfSchemaType = Exclude<JSONSchema, boolean>['type']
 /**
  * Defines the type of a value in the form that will be validated against the schema.
  */
-export type SchemaValue = string | number | ObjectValue | null | undefined | Array<SchemaValue> | boolean
+export type SchemaValue = string | number | ObjectValue | null | undefined | Array<SchemaValue> | boolean | File
 
 /**
  * A nested object value.
@@ -44,7 +44,14 @@ export interface JsonLogicRules {
     rule: RulesLogic
   }>
 }
-export interface JsonLogicRootSchema extends Pick<NonBooleanJsfSchema, 'if' | 'then' | 'else' | 'allOf' | 'anyOf' | 'oneOf' | 'not'> {}
+export type JsonLogicRootSchema = Pick<NonBooleanJsfSchema, 'if' | 'then' | 'else' | 'anyOf' | 'oneOf' | 'not'> & {
+  allOf?: (JsfSchema & { if?: JsonLogicIfNodeSchema })[]
+}
+
+export type JsonLogicIfNodeSchema = JsfSchema & {
+  validations?: Record<string, JsfSchema>
+  computedValues?: Record<string, JsfSchema>
+}
 
 export interface JsonLogicSchema extends JsonLogicRules, JsonLogicRootSchema {}